+++ nmostesting/NMOSTesting.py
@@ -83,7 +83,6 @@ from .suites import IS0901Test
 from .suites import IS0902Test
 # from .suites import IS1001Test
 from .suites import IS1101Test
-from .suites import IS1102Test
 from .suites import BCP00301Test
 from .suites import BCP0050101Test
 from .suites import BCP0060101Test
@@ -351,17 +350,6 @@ TEST_DEFINITIONS = {
         }],
         "class": IS1101Test.IS1101Test
     },
-    "IS-11-02": {
-        "name": "IS-11 Interaction with IS-04",
-        "specs": [{
-            "spec_key": "is-04",
-            "api_key": "node"
-        }, {
-            "spec_key": "is-11",
-            "api_key": "streamcompatibility"
-        }],
-        "class": IS1102Test.IS1102Test
-    },
     "BCP-003-01": {
         "name": "BCP-003-01 Secure Communication",
         "specs": [{
@@ -380,6 +368,10 @@ TEST_DEFINITIONS = {
     },
     "BCP-006-01-01": {
         "name": "BCP-006-01 NMOS With JPEG XS",
+        "specs": [{
+            "spec_key": "is-04",
+            "api_key": "node"
+        }],
         "extra_specs": [{
             "spec_key": "nmos-parameter-registers",
             "api_key": "flow-register"
