+++ nmostesting/suites/IS0501Test.py
@@ -1026,7 +1026,7 @@ class IS0501Test(GenericTest):
                 url = self.url + "single/senders/{}/transportfile".format(sender)
                 valid, response = self.do_request("GET", url)
                 if valid and response.status_code == 200:
-                    valid, message = check_content_type(response.headers, "application/sdp")
+                    valid, message = check_content_type(response.headers, ["application/sdp"])
                     if valid and message != "":
                         return test.FAIL(message)
                     elif not valid:
